Now supports decent concurrency when downloading your content. It is built for the following 3 scenarios: 1. lenOfData = 3000 e.g. mass scraping image boards 2. lenOfData = 1 URLs = 200 e.g. doujin 3. lenOfData = 1-10 but big file size e.g.hentai video here in main we will deal with the first type
THE INPUT FLAG ALSO CHANGED FROM -t TO -w

Please keep in mind to not over do it since some sites don't support it or will block you.
If you want to download something that does not support concurrent downloading, just don't supply the -w flag then it will be automatically be set to 1 and everything will be downloaded normally.

Functions used for M3U Download are now split into a seperate file.

Minor fixes:
A small bug in the hitomi.la extractor which was not able to download "manga" urls. This is now fixed.
Cleaned up the file sanitazion
If you use the -p flag the files will now end with the correct page numbers

New sites:
hentai2read.com